Move guilded cog under cogs
