+++ cogs/base/guildedcog.py
